fix(web): export computeInvoiceTax from contracts for client build
